I have fixed the Fisher-KPP, changed the rates
The rates were wrong, The production rate is no longer (per compartment) as its one d. Also degradation rate is going to be over h, since it is a 2nd order reaction. Note this is very slow as the number of PDE points increases - Must profile the code to understand whats taking so long.